Move operator library header out of the public API header files set.
