+++ src/operatorlibrary.h
